requirements.txt

torch>=2.2.0
torchvision>=0.17.0
torchaudio>=2.2.0 ; platform_system != "Windows"
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.3.0
Pillow>=10.0.0
opencv-python>=4.9.0.80
matplotlib>=3.8.0
tqdm>=4.66.0
streamlit>=1.35.0
typer>=0.12.0
pydantic>=2.6.0
pytest>=7.4.0
